move pii_like_keypaths into spec helper
